separate positive and negative eta outputs
In this scheme, the iEta is always an unsigned integer. For negative eta clusters (written out to the _negEta vector/link), iEta = 0 corresponds to the crystal closest to real eta = 0, and maximum iEta corresponds to the crystal at real eta -1.4841. For positive eta clusters (all written out to _posEta vectors/links), iEta = 0 represents the crystal closest to real eta = 0, and iEta maximum corresponds to the crystal at maximum real eta +1.4841.